requirements.txt
pytesseract
pdf2image
Pillow
pandas
openpyxl
streamlit
